Allow any context name
Before, using names that already existed as functions/constants/etc.
in the module would break due to the way that we checked for already
existing contexts. Now, the names given by the user are just keys in a
dictionary instead of actual type names inserted into the module (and
the type names are always safely created with `gensym`).