Record wrapped functions instead of keyword wrappers when possible
Previously, if `f(; x)` called g(x), then the `current_function` would
be something like `getfield(Main, Symbol("#kw##f"))` because of the
way that keyword functions are implemented. Now, it is simply `f` as
expected.
The only time that this seems to not work is when `f` is a closure.